Fix GstClockTime to UInt64 conversion in Query.swift
GstClockTime is UInt on Linux but UInt64 on macOS. Explicit
conversion ensures cross-platform compatibility.

+++ Sources/GStreamer/Query.swift
@@ -74,8 +74,8 @@ extension Pipeline {
 
         return LatencyInfo(
             isLive: live != 0,
-            minLatency: minLatency,
-            maxLatency: maxLatency
+            minLatency: UInt64(minLatency),
+            maxLatency: UInt64(maxLatency)
         )
     }
